added terrain class unfinished

+++ Engine/IndexedLineList.h
@@ -6,6 +6,5 @@
 struct IndexedLineList
 {
 	std::vector<Vec4> vertices;
-	std::vector<unsigned int> indices;
-	//std::vector<size_t> indices;
+	std::vector<size_t> indices;
 };
